+++ changelog.md
@@ -1,6 +1,9 @@
 
 
 # Changelog
+* 2019-04-24
+  * Added the Vertex.randomVertex(ViewPort) fuction.
+  * Added the VertexAnimator for demos.
 * 2019-04-17
   * Added the PlotBoilerplate.removeVertex(Vertex,boolean) function.
 * 2019-04-16